Add api to get context id in hw context class

+++ src/runtime_src/core/common/api/xrt_device.cpp
@@ -384,11 +384,11 @@ namespace xrt::aie {
 // of that partition. Get patition info using the context id passed and
 // convert relative column index to absolute using this info
 static uint16_t
-get_abs_col(const xrt_core::device* device, uint16_t context_id, uint16_t col)
+get_abs_col(const xrt_core::device* device, uint32_t context_id, uint16_t col)
 {
   auto data = xrt_core::device_query_default<xrt_core::query::aie_partition_info>(device, {});
   for (const auto& entry : data) {
-    if (std::stoi(entry.metadata.id) != context_id)
+    if (std::stoul(entry.metadata.id) != context_id)
       continue;
 
     auto abs_col = col + entry.start_col;
@@ -417,7 +417,7 @@ open_context(xrt::aie::device::access_mode am)
 
 std::vector<char>
 device::
-read_aie_mem(uint16_t context_id, uint16_t col, uint16_t row, uint32_t offset, uint32_t size) const
+read_aie_mem(uint32_t context_id, uint16_t col, uint16_t row, uint32_t offset, uint32_t size) const
 {
   return xdp::native::profiling_wrapper("xrt::aie::device::read_aie_mem",
     [this, &col, row, offset, size, context_id] {
@@ -434,7 +434,7 @@ read_aie_mem(uint16_t context_id, uint16_t col, uint16_t row, uint32_t offset, u
 
 size_t
 device::
-write_aie_mem(uint16_t context_id, uint16_t col, uint16_t row, uint32_t offset, const std::vector<char>& data)
+write_aie_mem(uint32_t context_id, uint16_t col, uint16_t row, uint32_t offset, const std::vector<char>& data)
 {
   return xdp::native::profiling_wrapper("xrt::aie::device::write_aie_mem",
     [this, &col, row, offset, &data, context_id] {
@@ -450,7 +450,7 @@ write_aie_mem(uint16_t context_id, uint16_t col, uint16_t row, uint32_t offset, 
 }
 uint32_t
 device::
-read_aie_reg(uint16_t context_id, uint16_t col, uint16_t row, uint32_t reg_addr) const
+read_aie_reg(uint32_t context_id, uint16_t col, uint16_t row, uint32_t reg_addr) const
 {
   return xdp::native::profiling_wrapper("xrt::device::read_aie_reg",
     [this, &col, row, reg_addr, context_id] {
@@ -467,7 +467,7 @@ read_aie_reg(uint16_t context_id, uint16_t col, uint16_t row, uint32_t reg_addr)
 
 bool
 device::
-write_aie_reg(uint16_t context_id, uint16_t col, uint16_t row, uint32_t reg_addr, uint32_t reg_val)
+write_aie_reg(uint32_t context_id, uint16_t col, uint16_t row, uint32_t reg_addr, uint32_t reg_val)
 {
   return xdp::native::profiling_wrapper("xrt::device::write_aie_reg",
     [this, &col, row, reg_addr, &reg_val, context_id] {
